pac_issues_analysis/scrape_goog_web_urls/requirements.txt
pandas
googlesearch_python==2020.0.2
bs4
html5lib
lxml